Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-position/position-sticky-transforms-translate-expected.txt


PASS Translation transform can move sticky element past sticking point
PASS Stuck elements can still be moved via translations
FAIL The sticky element should stick before the container is offset by a translation assert_equals: expected 150 but got 100
